Uses SRP to separate some methods of TextUi into Formatter class
